requirements.txt
sentence-transformers
psycopg2-binary
pgvector
python-dotenv
openai
pypdf
python-docx
tqdm
numpy
